Minor: Use primitives instead of boxed numbers
Don't use boxed primitives when non-nullness is guaranteed.

Also some timestamps changed to use -1 as a "unset" value.
No changes to ServiceDocuments.

+++ xenon-common/src/test/java/com/vmware/xenon/common/TestOperation.java
@@ -688,7 +688,7 @@ public class TestOperation extends BasicReusableHostTestCase {
         assertEquals(op.getUri().getPort(), sop.port);
         assertEquals(op.getUri().getPath(), sop.path);
         assertEquals(op.getUri().getQuery(), sop.query);
-        assertEquals(op.getId(), sop.id.longValue());
+        assertEquals(op.getId(), sop.id);
         assertEquals(op.getReferer(), sop.referer);
         assertEquals(op.getBodyRaw(), sop.jsonBody);
         assertEquals(op.getStatusCode(), sop.statusCode);
